Implement primitive worker

+++ src/primitive_request.rs
@@ -3,9 +3,9 @@ use std::path::PathBuf;
 pub struct PrimitiveRequest {
     pub request_id: String,
     pub input_file_path: PathBuf,
-    pub num_shapes: u64,
-    pub max_age: u64,
-    pub scale_to: u64,
-    pub seed: u64,
+    pub num_shapes: u32,
+    pub max_age: u32,
+    pub scale_to: u32,
+    pub seed: u32,
     pub shape: String
 }
